Explicitly cast to size_t to avoid overflows in pointer arithmetic.
This is more self-documenting than the multiply() utility we had before.
We also use an overflow-resistant calculation for the chunk cache size.

+++ include/tatami_hdf5/load_hdf5_matrix.hpp
@@ -112,7 +112,7 @@ tatami::DenseMatrix<!transpose_, Value_, Index_, ValueStorage_> load_hdf5_dense_
     auto dhandle = open_and_check_dataset<false>(fhandle, name);
 
     auto dims = get_array_dimensions<2>(dhandle, name);
-    ValueStorage_ values(dims[0] * dims[1]);
+    ValueStorage_ values(static_cast<size_t>(dims[0]) * static_cast<size_t>(dims[1])); // cast just in case hsize_t is something silly...
     dhandle.read(values.data(), define_mem_type<Stored<ValueStorage_> >());
 
     if constexpr(transpose_) {
